[MINOR] Fixing some functional tests by moving to right packages (#3596)
